arm: dts: aspeed: tyan s8036: Enable kcs interrupts
When the BIOS is built with kcs interrupts enabled, not enabling
interrupts on the BMC results in very poor IPMI performance.
The other way around (BIOS with interrupts disabled, BMC with
interrupts enabled) doesn't suffer degraded IPMI performance.
Enabling interrupts on the BMC covers both scenarios, and should
be the default.

TESTED: manually verified IPMI performance when BIOS is built with and
without KCS interrupts.

Signed-off-by: Ali El-Haj-Mahmoud <aaelhaj@google.com>
Link: https://lore.kernel.org/r/20230118150030.2079226-1-aaelhaj@google.com
Signed-off-by: Joel Stanley <joel@jms.id.au>

[ upstream commit: 8803d9438ef65c96b03ae95472e19b3ac072c930 ]