arm64: ls1046ardb: Use in-band-status for SFP module
net10 is connected to an SFP module. Unfortunately, the I2C lines are
not connected due to an address conflict. Now that DPAA uses phylink, we
can use in-band-status. This lets us determine whether the link is up or
down instead of assuming it is up all the time. Also fix the phy mode
while we're here.

Signed-off-by: Sean Anderson <sean.anderson@seco.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 46e828026cc8daf29b251efca27b93eaaf906081 ]