ARM: dts: exynos: move exynos-bus nodes out of soc in Exynos4210
The soc node is supposed to have only device nodes with MMIO addresses,
as reported by dtc W=1:

  exynos4210.dtsi:218.20-224.5:
    Warning (simple_bus_reg): /soc/bus-dmc: missing or empty reg/ranges property

and dtbs_check:

  exynos4210-i9100.dtb: soc: bus-dmc:
    {'compatible': ['samsung,exynos-bus'], 'clocks': [[5, 457]], 'clock-names': ['bus'], 'operating-points-v2': [[82]], 'status': ['disabled']} should not be valid under {'type': 'object'}

Move the bus nodes and their OPP tables out of SoC to fix this.
Re-order them alphabetically while moving and put some of the OPP tables
in device nodes (if they are not shared).

Link: https://lore.kernel.org/r/20230125094513.155063-4-krzysztof.kozlowski@linaro.org
Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>

[ upstream commit: 09dd37396ab6e7d1ce66a7cb22019cea19c1d338 ]