dt-bindings: watchdog: migrate rt2880 text bindings to YAML
Ralink RT2880 Watchdog bindings used text format, so migrate them to YAML.
There are some additions to the binding that were not in the original
txt file. This binding is used in RT2880, RT3050, RT3352, RT3883, RT5350,
and MT7620 SoCs. To properly align binding with driver code we need to add
to the schema 'clocks' and 'resets' properties.

Signed-off-by: Sergio Paracuellos <sergio.paracuellos@gmail.com>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Link: https://lore.kernel.org/r/20230302085914.2858645-1-sergio.paracuellos@gmail.com
Signed-off-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Wim Van Sebroeck <wim@linux-watchdog.org>

[ upstream commit: fe5631bb3a4bf62208ed52600c21dd3d1787f412 ]