dt-bindings: remoteproc: qcom: sm6115-pas: Add QCM2290
QCM2290 is more or less a fork of the same design that SM6115 is based
on. As a result, the ADSP and modem found on it are identical.

Add compatibles for the QCM2290 with SM6115 fallbacks so as not to
require any driver changes. Change the allOf:if:properties clauses
to look for the presence of SM6115 compatibles and not an exact match.

Note that QCM2290 lacks a CDSP.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/20230315183231.3562580-1-konrad.dybcio@linaro.org
Signed-off-by: Bjorn Andersson <andersson@kernel.org>

[ upstream commit: b352033e19c4591df299a8f623508c5c2ce4c5b3 ]