ARM: dts: qcom: apq8074-dragonboard: enable adsp and MSS
Enable ADSP and Modem DSPs on APQ8074 dragonboard. The MSS region
differs from the one defined in the msm8974, so it overriden locally.

The modem is specified use mba.mbn instead of mbn.b00 (for the sake of
similarity with other platforms). This requires a patch for remoteproc
to be applied [1].

[1] https://lore.kernel.org/all/20230507172041.2320279-1-dmitry.baryshkov@linaro.org/

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Reviewed-by: Konrad Dybcio <konrad.dybcio@linaro.org>
Signed-off-by: Bjorn Andersson <andersson@kernel.org>
Link: https://lore.kernel.org/r/20230507190735.2333145-4-dmitry.baryshkov@linaro.org

[ upstream commit: 1ca793c3714a754488c782eb0d68b279d811b977 ]