dt-bindings: soc: bcm: add missing dt-binding document for bcm23550-cdc
Broadcom BCM23550 has a Cluster Dormant Control block that keeps the CPU
in idle state. A command needs to be sent to this block to bring the CPU
into running state.

This has been in use in mainline Linux since ~2016, but was never
documented. Add a dt-binding document for it.

Signed-off-by: Stanislav Jakubek <stano.jakubek@gmail.com>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/20230508135930.GA27583@standask-GA-A55M-S2HP
Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>

[ upstream commit: 19571b46d8997d20c10b7194a3ee5772500ce725 ]