dt-bindings: backlight: pwm: Make power-supply not required
make dtbs_check:

    arch/arm/boot/dts/renesas/r8a7742-iwg21d-q7.dtb: backlight: 'power-supply' is a required property
	    From schema: Documentation/devicetree/bindings/leds/backlight/pwm-backlight.yaml

As that backlight device node already has an "enable-gpios" property to
control the power supplied to the backlight, it sounds a bit silly to
have to add a "power-supply" property just to silence this warning.  In
addition, as of commit deaeeda2051fa280 ("backlight: pwm_bl: Don't rely
on a disabled PWM emiting inactive state"), the Linux driver considers
the power supply optional.

Fix this by synchronizing the bindings with actual driver behavior by
making the "power-supply" optional.

Fixes: deaeeda2051fa280 ("backlight: pwm_bl: Don't rely on a disabled PWM emiting inactive state")
Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Acked-by: Conor Dooley <conor.dooley@microchip.com>
Link: https://lore.kernel.org/r/29943059c80c8db0db437f9548f084a67326647b.1684856131.git.geert+renesas@glider.be
Signed-off-by: Lee Jones <lee@kernel.org>

[ upstream commit: a8b09e6fe3eee4525ef6b6d5bd0e69c785e2f757 ]