ARM: dts: imx6dl: lanmcu: Disable unused USB PHY nodes
In the imx6dl-lanmcu system, the usbh2 and usbh3 interfaces are not in
use. However, the related USB PHY nodes (usbphynop1 and usbphynop2) are
enabled by default in the imx6qdl.dtsi file, which can lead to confusion
and potential resource mismanagement.

This commit explicitly disables these unused USB PHY nodes in the
imx6dl-lanmcu device tree, aligning the configuration more accurately
with the actual hardware setup.

Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: bcadb701689a6569fbec7e60e21a73c586d0ed3b ]