dt-bindings: mfd: maxim,max77693: Add USB connector
Add micro-USB connector under "muic" node to properly represent the
hardware.  Deprecate also the old "max77693-muic" in favor of generic
"muic" (this is max77693, so there is no need to state it in its child
nodes).  This "muic" node is used only to instantiate MUIC driver by
compatible, thus actual Linux driver changes are needed.  Extend the
example with this new code.

Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Reviewed-by: Rob Herring <robh@kernel.org>
Reviewed-by: Chanwoo Choi <cw00.choi@samsung.com>
Link: https://lore.kernel.org/r/20230723142842.98204-1-krzysztof.kozlowski@linaro.org
Signed-off-by: Lee Jones <lee@kernel.org>

[ upstream commit: 789c9ce9b46facdf405572b4d0387ed67a49a97f ]